+++ integration-test-jdbc/src/test/kotlin/integration/jdbc/SqlSelectQueryForUpdateTest.kt
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
 import org.komapper.core.dsl.SqlDsl
-import org.komapper.core.dsl.desc
+import org.komapper.core.dsl.operator.desc
 import org.komapper.jdbc.JdbcDatabase
 
 @ExtendWith(Env::class)
