+++ integration-test-jdbc/src/test/kotlin/integration/jdbc/SqlSelectQuerySubqueryTest.kt
@@ -7,8 +7,8 @@ import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
 import org.komapper.core.dsl.SqlDsl
-import org.komapper.core.dsl.count
-import org.komapper.core.dsl.max
+import org.komapper.core.dsl.operator.count
+import org.komapper.core.dsl.operator.max
 import org.komapper.jdbc.JdbcDatabase
 
 @ExtendWith(Env::class)
