+++ integration-test-jdbc/src/test/kotlin/integration/jdbc/ValueClassTest.kt
@@ -17,12 +17,12 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
 import org.komapper.core.dsl.EntityDsl
 import org.komapper.core.dsl.SqlDsl
-import org.komapper.core.dsl.case
-import org.komapper.core.dsl.concat
-import org.komapper.core.dsl.count
 import org.komapper.core.dsl.expression.When
-import org.komapper.core.dsl.max
-import org.komapper.core.dsl.plus
+import org.komapper.core.dsl.operator.case
+import org.komapper.core.dsl.operator.concat
+import org.komapper.core.dsl.operator.count
+import org.komapper.core.dsl.operator.max
+import org.komapper.core.dsl.operator.plus
 import org.komapper.jdbc.JdbcDatabase
 
 @ExtendWith(Env::class)
