+++ integration-test-r2dbc/src/test/kotlin/integration/r2dbc/EntitySelectQueryTest.kt
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.Assertions.assertNull
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
 import org.komapper.core.dsl.EntityDsl
-import org.komapper.core.dsl.desc
+import org.komapper.core.dsl.operator.desc
 import org.komapper.r2dbc.R2dbcDatabase
 
 @ExtendWith(Env::class)
