+++ integration-test-r2dbc/src/test/kotlin/integration/r2dbc/SqlSelectQuerySubqueryTest.kt
@@ -7,8 +7,8 @@ import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
 import org.komapper.core.dsl.SqlDsl
-import org.komapper.core.dsl.count
-import org.komapper.core.dsl.max
+import org.komapper.core.dsl.operator.count
+import org.komapper.core.dsl.operator.max
 import org.komapper.r2dbc.R2dbcDatabase
 
 @ExtendWith(Env::class)
