+++ komapper-core/src/main/kotlin/org/komapper/core/dsl/operator/aliasExpression.kt
@@ -1,0 +1,8 @@
+package org.komapper.core.dsl.operator
+
+import org.komapper.core.dsl.expression.AliasExpression
+import org.komapper.core.dsl.expression.ColumnExpression
+
+infix fun <T : Any, S : Any> ColumnExpression<T, S>.alias(alias: String): ColumnExpression<T, S> {
+    return AliasExpression(this, alias)
+}
